Final Set of Features implemented
Adding active breaking, boost, network channel selection

+++ KickerRobot/main/drivers/servo.cpp
@@ -24,11 +24,11 @@ void servo::setDegree(uint16_t target_angle){
     }
     uint16_t pulse_width = 0;
     if(direction){
-        pulse_width = pwm_low_limit + (target_angle * (pwm_high_limit-pwm_low_limit)/max_servo_angle) + offset;
+        pulse_width = pwm_low_limit + (target_angle * (pwm_high_limit-pwm_low_limit)/(float)max_servo_angle) + offset;
     }else{
-        pulse_width = pwm_high_limit - (target_angle * (pwm_high_limit-pwm_low_limit)/max_servo_angle) + offset;
+        pulse_width = pwm_high_limit - (target_angle * (pwm_high_limit-pwm_low_limit)/(float)max_servo_angle) + offset;
     }
-
+    //printf("%d\n", pulse_width);
     uint32_t duty = (pulse_width * (1 << LEDC_TIMER_12_BIT)) / servo_period_time; //20,000 us is the period
     (void)ledc_set_duty(LEDC_LOW_SPEED_MODE, ledc_timer, duty);
     (void)ledc_update_duty(LEDC_LOW_SPEED_MODE, ledc_timer);
